// frontend/src/types.ts
export interface UserType {
    anonymize_data: boolean
    distinct_id: string
    email: string
    email_opt_in: boolean
    id: number
    name: string
    // posthog_version: string
    organization: OrganizationType | null
    team: TeamType | null
    toolbar_mode: 'disabled' | 'toolbar'
    organizations: OrganizationType[]
    teams: TeamType[]
    current_organization_id: string
    current_team_id: number
    plugin_access: PluginAccess
    has_password: boolean
    is_multi_tenancy: boolean
    is_staff: boolean
    is_debug: boolean
    is_impersonated: boolean
    ee_enabled: boolean
    email_service_available: boolean
    realm: 'cloud' | 'hosted'
    billing?: OrganizationBilling
}

/* Type for User objects in nested serializers (e.g. created_by) */
export interface UserNestedType {
    id: number
    distinct_id: string
    first_name: string
    email: string
}

export interface UserUpdateType {
    user?: Omit<Partial<UserType>, 'team'>
    team?: Partial<TeamType>
}

export interface PluginAccess {
    view: boolean
    install: boolean
    configure: boolean
}

export interface PersonalAPIKeyType {
    id: string
    label: string
    value?: string
    created_at: string
    last_used_at: string
    team_id: number
    user_id: string
}

export interface OrganizationType {
    id: string
    name: string
    created_at: string
    updated_at: boolean
    available_features: string[]
    billing_plan: string
    billing: OrganizationBilling
    teams?: TeamType[]
    membership_level: OrganizationMembershipLevel | null
    setup: SetupState
    personalization: PersonalizationData
}

export interface OrganizationMemberType {
    joined_at: string
    level: OrganizationMembershipLevel
    membership_id: string
    updated_at: string
    user_email: string
    user_first_name: string
    user_id: number
}

export interface EventUsageType {
    event: string
    usage_count: number
    volume: number
}

export interface PropertyUsageType {
    key: string
    usage_count: number
    volume: number
}

export interface TeamType {
    id: number
    name: string
    anonymize_ips: boolean
    api_token: string
    app_urls: string[]
    completed_snippet_onboarding: boolean
    event_names: string[]
    event_properties: string[]
    event_properties_numerical: string[]
    event_names_with_usage: EventUsageType[]
    event_properties_with_usage: PropertyUsageType[]
    slack_incoming_webhook: string
    session_recording_opt_in: boolean
    session_recording_retention_period_days: number | null
    plugins_opt_in: boolean
    ingested_event: boolean
    is_demo: boolean
}
// ...
export interface FormattedNumber {
    // :TODO: DEPRECATED, formatting will now happen client-side
    value: number
    formatted: string
}

export interface OrganizationBilling {
    plan: PlanInterface | null
    current_usage: FormattedNumber | number | null
    should_setup_billing?: boolean
    stripe_checkout_session?: string
    subscription_url?: string
    event_allocation: FormattedNumber | number | null
}

export interface PlanInterface {
    key: string
    name: string
    custom_setup_billing_message: string
    image_url: string
    self_serve: boolean
    is_metered_billing: boolean
    allowance: FormattedNumber | number | null // :TODO: DEPRECATED
    event_allowance: number
    price_string: string
}
// ...
export type PersonalizationData = Record<string, string | string[] | null>

interface EnabledSetupState {
    is_active: true // Whether the onbarding setup is currently active
    current_section: number
    any_project_ingested_events: boolean
    any_project_completed_snippet_onboarding: boolean
    non_demo_team_id: number | null
    has_invited_team_members: boolean
}

interface DisabledSetupState {
    is_active: false
    current_section: null
}

export type SetupState = EnabledSetupState | DisabledSetupState